Golistica-main/futbol5_project/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
bcrypt
passlib[bcrypt]
python-jose[cryptography]
cryptography
pydantic
python-dotenv
python-multipart
